requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cassandra-driver==3.29.1
certifi==2024.2.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kafka-python==2.0.2
openai==1.17.1
playwright==1.43.0
py4j==0.10.9.7
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
pyspark==3.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
